Start on Server code Refactor to provide better structure and implement new lobby
Refactor extra server functions and code not related to client and server calls into a gameWorld class.

Implement a lobby class to better handle code related to adding players to a lobby and handling disconnects.

Implement a user class to handle client variables that are not necessary in the gameworld entities players objects.