Modify Lobby to start filled with 4 AI players
When clients connected they now replace one of the 4 AI controlled players.
The game now only starts if the minimum connected players is met.
Player positions are now set after the AI players are created to give them proper starting positions.
Fixed a bug with the AI update timer that stopped the timer after the first interval.
Starting to test the AI logic