Allow AI controlled players to check for entity and tile collisions
beautified.
allowed for choose hero to choose between 1 and 4 instead of 1 and 2.
added a null check to the gameWorld check collisions method so the AI controlled players can still use it without an attached client.
implement a move method to the AI similar to the clients.
started to allow ai to check distances to other entities - INCOMPLETE