Modify game timer and fix game reset issues
fixed an issue where if a player had joined a game and the server then restarted, when they joined a new game there was a change their lobby still had the old sprites and multiple frogs.

Fixed an issue where  a timer interval not quite being cancelled fast enough on gameover caused a crash when the interval tried to move the powerup.
try to fix issues with allowing players to start a new game from gameover ( seems tempermental)